Update css-layouts.md
Github fix spelling challenge. Idk if this is what I'm supposed to do but I looked through prior lessons and this was the first thing I found.

Line 142 has a grammar error. It is written "+ **Device Screen Size:** Small screens cannot have the same number of columns as larger ones; the images may need to be smaller, among other considerations.  E.g:"
There should be a period after the g in E.g.: